backend/db_test.py
import psycopg2
import os
from dotenv import load_dotenv

load_dotenv()

try:
    conn = psycopg2.connect(
        host="127.0.0.1",
        database="fastapi_db",
        user="user",
        password="password",
        port="5433"
    )
    print("Connection successful!")
    conn.close()
except Exception as e:
    print(f"Connection failed: {e}")
